DHFPROD-5887: Removing as much state as possible from HubTestBase
The intent is for tests to use getHubClient() whenever possible to get a DatabaseClient based on the current user. 

Had to make one app change - LegacyCollectorImpl now only specifies a database if the DatabaseClient specifies one. 

Moved helper methods in HubTestBase that were only used in 1 or 2 tests into those tests themselves, just to slim down the number of methods for developers to look at for reuse. 

Also removed ssl/cert support from tests, as that appears unused. Can add it back later if needed, but would prefer to do so in a way involving a lot less code.